Fixed NullPointerException on javaProcess in JarController.stopService.

Former-commit-id: 54683674180cd3ea252dec8b340c533f9a1ffc51

+++ ServiceNode/src/main/java/ie/ucd/dempsey/mecframework/service/JarController.java
@@ -84,6 +84,8 @@ public class JarController implements ServiceController {
 
     @Override
     public void stopService() {
+        if (isServiceRunning()) return;
+
         logger.info("Trying to stop the Java process. javaProcess.isAlive?={}", javaProcess.isAlive());
         javaProcess.destroyForcibly();
         try {
